Add django-reversion integration to support page versioning
Add support for versioning of Fluent and Flat page content with the
django-reversion library.

If django-reversion is installed and the new app setting
FLUENT_PAGES_DISABLE_REVERSION is falsey, the default admins for fluent
and flat pages are extended with reversion support so these page types
can be versioned, reverted, and restored in the Django admin.

WARNING: This implementation is a work in progress and is currently
buggy and incomplete, but the basics are mostly working and things will
improve.

IMPLEMENTATION NOTES:

 * This currently requires a forked version of django-reversion
   available as the master branch of github.com/ixc/django-reversion

 * New strict-restore support in fork of django-reversion should be
   enabled by setting REVERSION_ADMIN_STRICT_REVERT
   and REVERSION_ADMIN_STRICT_REVERT_WITH_PRE_DELETE to True

 * FLUENT_PAGES_PARENT_ADMIN_MIXIN must be set to:
   'fluent_pages.integration.django_reversion.admin_mixins.FluentReversionAdminMixin'

 * If Haystack is configured to perform realtime updates, to avoid
   errors when indexing reverted/restored pages the
   HAYSTACK_SIGNAL_PROCESSOR setting should be set to:
   'fluent_pages.integration.django_reversion.haystack_utils.ReversionRealtimeSignalProcessor'