More generic roll-up of version data into object form data
Make logic to recognise and handle special `content` and `layout` forms
in Flat/FluentPage models much more generic, rather than requiring the
model to match expected ContentType names.

This means that the logic should now work for arbitrary page subclasses,
not just Fluent's own FlatPage and FluentPage.

There is a risk that this is too generic...