Add versioning admin mixin for polymorphic parents
Add a new class ReversionFluentContentsParentPageAdminMixin to be used
as a basis for polymorphic parent admins that do not extend the full
default Flat or Fluent page admins. This new mixin does a little extra
work to ensure that the appropriate child admin for a polymorphic model
is used when applying changes in the admin, instead of the parent admin.