Trigger pre-revert/recover signal within reversion's DB transaction
Send the the `pre_revisionform_revert` signal from *within* the atomic
DB transaction created inside `VersionAdmin.revisionform_view` so that
any custom pre-processing we do is applied in the DB only when the
revision form is submitted, not when it is only viewed pre-submission.

To make this happen we need to reinstate the method override hack we
used previously, to delay sending the signal until reversion's internal
`revert` method is called.

We could avoid this hack by forking django-reversion and putting the
new signal send calls in the appropriate place there, instead of in our
django-fluent-pages fork, but I think avoiding yet another fork is
worth the trade-off of this ugly approach, which we have been doing for
a while anyway.